requirements.txt
sentence-transformers
faiss-cpu
transformers
fastapi
uvicorn
scikit-learn
hdbscan
requests
